Implement VBA component import code
Implement code to import VBA code into a workbook (code needs some cleanup)
Refactor how the importer and exporter work to take the directory to target
in the actual import/export function instead of when getting a class off the
API. This gives the caller greater flexibility to run multiple imports from
multiple folders. This was required to support functionality to delete all
modules that were not imported (mirror data from folders).

Issue #1.