fix: handle deletion of sheet components (#9)
Update the code deleting any components that were not imported to better
handle sheet components. Instead of trying to remove the sheet component it
instead removes all the code stored inside that component. This is required
because Excel does not allow deleting a sheet component without deleting
the corresponding worksheet.